+++ src/test/java/de/codescape/jira/plugins/scrumpoker/model/AdditionalFieldTest.java
@@ -3,9 +3,11 @@ package de.codescape.jira.plugins.scrumpoker.model;
 import com.atlassian.jira.issue.fields.CustomField;
 import org.junit.Test;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
 import static org.mockito.Mockito.mock;
 
 public class AdditionalFieldTest {