+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/EstimateFieldServiceImplTest.java
@@ -31,7 +31,6 @@ import static de.codescape.jira.plugins.scrumpoker.service.EstimateFieldServiceI
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
 public class EstimateFieldServiceImplTest {