+++ src/test/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerLicenseServiceImplTest.java
@@ -13,8 +13,8 @@ import org.mockito.junit.MockitoRule;
 
 import static de.codescape.jira.plugins.scrumpoker.service.ScrumPokerLicenseServiceImpl.ERROR_MESSAGE_PREFIX;
 import static de.codescape.jira.plugins.scrumpoker.service.ScrumPokerLicenseServiceImpl.MISSING_LICENSE;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;