+++ src/test/java/de/codescape/jira/plugins/scrumpoker/workflow/StartSessionWorkflowFunctionTest.java
@@ -16,8 +16,6 @@ import org.mockito.junit.MockitoRule;
 
 import java.util.Map;
 
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
 public class StartSessionWorkflowFunctionTest {